Malaria Detection in Blood Samples Using Convolutional Neural Networks
Malaria is the leading cause of morbidity and mortality in tropical and subtropical countries. WHO
estimates the number of malaria deaths at 435,000 for 2018. Machine learning has great potential
to lighten the burden of malaria in temperate regions around the world where mosquitoes thrive
especially in remote sub-Saharan Africa.

However, the accuracy of the manual method of malaria
diagnosis using microscopy depends on the human expert. However, it is prone to some
shortcomings which include time consumption and excessive workload for the pathologists. We
propose an automated diagnostic system that can exclude the human expert from the process or
serve as an aid for an expert to lower workload and improve accuracy. The ultimate goal of this
work was to develop a system for detecting malaria using microscopic images of stained blood
samples.